feat(chain,core)!: move `Merge` to `bdk_core`
